# {Core} Stop importing pkg_resources when getting versions