… src/azure-cli/azure/cli/command_modules/util/tests/latest/test_version.py