## Description

`az version` and the **version check** introduced in #14803 query the local versions of Azure CLI modules:

- `azure-cli`
- `azure-cli-command-modules-nspkg`
- `azure-cli-core`
- `azure-cli-nspkg`
- `azure-cli-telemetry`

The current implementation uses `pkg_resources.working_set` which takes ~200ms to load:

```
python -X importtime -m azure.cli version 2>perf.log ; tuna .\perf.log
```

![image](https://user-images.githubusercontent.com/4003950/91160654-41e69900-e6fc-11ea-9f76-b4bc065fb6c6.png)

This is outstanding compared to `az version`'s 500ms running time.

This PR uses the **hard-coded version** instead of querying all modules under `site-packages` using `pkg_resources`.

### Removal of `azure-cli-dev-tools` and `azure-cli-testsdk`

This PR removes `azure-cli-dev-tools` and `azure-cli-testsdk`. The removal only applies to **dev CLI**, and won't affect the **released CLI** (they are not present at all). It is safe for public users. 

After the change, the dev CLI and released CLI have the same output.

```sh
# Before (dev)
> az version
{
  "azure-cli": "2.11.0",
  "azure-cli-core": "2.11.0",
  "azure-cli-dev-tools": "0.1.1",
  "azure-cli-telemetry": "1.0.5",
  "azure-cli-testsdk": "0.2.4",
  "extensions": {
    "interactive": "0.4.4",
    "log-analytics": "0.2.1",
    "portal": "0.1.1",
    "resource-graph": "1.1.0"
  }
}

# After (dev and release)
> az version
{
  "azure-cli": "2.11.0",
  "azure-cli-core": "2.11.0",
  "azure-cli-telemetry": "1.0.5",
  "extensions": {
    "interactive": "0.4.4",
    "log-analytics": "0.2.1",
    "portal": "0.1.1",
    "resource-graph": "1.1.0"
  }
}
```

### Performance

The performance improvement is significant - from ~500ms to ~300ms:

```sh
> Measure-Command {az version}
TotalMilliseconds : 531.9338

> Measure-Command {az version}
TotalMilliseconds : 358.5301
```

`site-packages` is not imported anymore:

![image](https://user-images.githubusercontent.com/4003950/91160768-6773a280-e6fc-11ea-9e95-ab9bd37a5774.png)

## Testing Guide

```sh
az version
Measure-Command {az version}
```